Made query panel look like user request panel. Also fixed the default sizing issue, at least on my computer.
Keep an eye on this, since I just found experimentally how large the layout makes the content and set the preferred size to that. If the layout sizing is computed dynamically somehow, perhaps based on available screen space, then this would not work. It *should* have just enough room for all the content (you can tell from the border around the main content), give or take a few pixels.